requirements.txt
yt-dlp>=2024.1.0
youtube-transcript-api>=0.6.0
sentence-transformers>=2.7.0
weasyprint>=61.0
jinja2>=3.1.0
pyyaml>=6.0
requests>=2.31.0
numpy>=1.26.0
